requirements.txt
asgiref==3.8.1
Django==3.2.5
djangorestframework==3.12.4
pytz==2024.1
sqlparse==0.5.0
